learning_module_generator/backend/requirements.txt
Django==4.2
requests==2.31.0
firebase-admin==6.5.0
gunicorn==20.1.0
python-dotenv==1.0.1
